+++ activerecord/lib/active_record/associations/builder/belongs_to.rb
@@ -34,7 +34,10 @@ module ActiveRecord::Associations::Builder
         method_name = "belongs_to_counter_cache_before_destroy_for_#{name}"
         mixin.redefine_method(method_name) do
           record = send(name)
-          record.class.decrement_counter(cache_column, record.id) unless record.nil?
+
+          if record && !self.destroyed?
+            record.class.decrement_counter(cache_column, record.id)
+          end
         end
         model.before_destroy(method_name)